add replace from apps to utils
